+++ internal/remote_control/remoteagent/card_metadata.go
