+++ internal/remote_control/remoteagent/command_adapter.go
