+++ internal/remote_control/remoteagent/executor_claude_permission_test.go
