+++ internal/remote_control/remoteagent/handler_send.go
@@ -5,10 +5,10 @@ import (
 	"strings"
 
 	"github.com/sirupsen/logrus"
+	"github.com/tingly-dev/tingly-box/internal/remote_control/feature"
 
 	"github.com/tingly-dev/tingly-box/imbot"
 	"github.com/tingly-dev/tingly-box/internal/remote_control/bot"
-	"github.com/tingly-dev/tingly-box/internal/remote_control/bot/feature"
 )
 
 func (h *BotHandler) SendText(hCtx HandlerContext, text string) {
