+++ internal/remote_control/remoteagent/stream.go
