+++ internal/remote_control/remoteagent/stream_aggregate_test.go
