+++ internal/remote_control/remoteagent/testharness.go
@@ -12,7 +12,7 @@ import (
 	"github.com/tingly-dev/tingly-box/imbot"
 	"github.com/tingly-dev/tingly-box/internal/data/db"
 	"github.com/tingly-dev/tingly-box/internal/remote_control/bot"
-	"github.com/tingly-dev/tingly-box/internal/remote_control/bot/feature"
+	"github.com/tingly-dev/tingly-box/internal/remote_control/feature"
 	"github.com/tingly-dev/tingly-box/remote/session"
 )
