+++ internal/remote_control/remoteagent/util_extra_test.go
